should not be patched for darwin

+++ default/scripts/cvc4.sh
@@ -1,6 +1,8 @@
 cd cvc4
 export MACHINE_TYPE=x86_64
-sed -i -re 's,rm -rf src/antlr3debughandlers.c \&\& touch src/antlr3debughandlers.c,rm -rf src/antlr3debughandlers.c \&\& touch src/antlr3debughandlers.c \&\& cp  /usr/share/misc/config.* . ,g' ./contrib/get-antlr-3.4
+if [ ${ARCH_BASE} != 'darwin' ]; then
+    sed -i -re 's,rm -rf src/antlr3debughandlers.c \&\& touch src/antlr3debughandlers.c,rm -rf src/antlr3debughandlers.c \&\& touch src/antlr3debughandlers.c \&\& cp  /usr/share/misc/config.* . ,g' ./contrib/get-antlr-3.4
+fi
 ANTLR_CONFIGURE_ARGS="--host=${CROSS_NAME} --build=`gcc -dumpmachine`"  ./contrib/get-antlr-3.4
 git clone https://github.com/uiri/toml.git
 export PYTHONPATH=$PYTHONPATH:`pwd`/toml
